requirements.txt
yfinance==0.2.18
psycopg2-binary==2.9.7
pandas==2.0.3
python-dotenv==1.0.0
requests==2.31.0
numpy==1.24.3
fuzzywuzzy==0.18.0
python-Levenshtein==0.21.1